requirements.txt
pandas>=1.3,<1.6
pytz>=2021.1
sqlalchemy>=1.4,<2.0
psycopg2-binary>=2.9
python-dotenv>=0.20
requests>=2.28

